make reset more secure

+++ reset_test_dir.sh
@@ -2,11 +2,17 @@
 
 dir=$1
 
-echo 
-
 shopt -s extglob
-cd $dir
-rm -r  !(meta|unmapped_bam) 2>/dev/null
-rm *.done 2>/dev/null
 
-echo "$dir is now reset."
+read -p "Do you really want to reset the directory $dir (y/n)? " RESP
+if [ $RESP = "y" ]; then
+  cd $dir
+  rm -r !(meta|unmapped_bam) 2>/dev/null
+  rm *.done 2>/dev/null
+  echo "$dir is now reset."
+elif [ $RESP = "n" ]; then
+  echo "Abort!"
+else
+  echo "Unknown entry.(y/n) needed.Exiting!"
+fi 
+
